chore: rename `prisma.js` to `dbClient.js`
