+++ src/hooks.server.js
@@ -14,7 +14,7 @@ import { settings }  from "$lib/settings.js"
 // #region Handles
     // #region auth()
         // #region Specific Imports
-import dbClient from "$lib/server/database/prisma/prisma.js"
+import dbClient from "$lib/server/database/prisma/dbClient.js"
 import inputHandler from "$lib/server/utils/inputHandler.js"
 import logError from "$lib/server/utils/errorLogger.js"
 import { dateFromNow } from "$lib/client/utils/dateUtils.js"
