+++ src/routes/(main)/(private)/(user)/friends/+page.server.js
@@ -121,7 +121,7 @@ export const load = async ({ locals }) => {
 
 // #region actions
     // #region Specific Imports
-import dbClient from "$lib/server/database/prisma/prisma.js"
+import dbClient from "$lib/server/database/prisma/dbClient.js"
 import logError from "$lib/server/utils/errorLogger.js"
     // #endregion
