+++ src/routes/(main)/(private)/(user)/settings/+page.server.js
@@ -1,5 +1,5 @@
 // #region Imports
-import dbClient from "$lib/server/database/prisma/prisma.js"
+import dbClient from "$lib/server/database/prisma/dbClient.js"
 import inputHandler from "$lib/server/utils/inputHandler.js"
 import { stringHasher } from "$lib/server/utils/hashUtils.js"
 import { emailer } from "$lib/server/utils/emailUtils.js"
