+++ src/routes/(main)/(private)/logout/+page.server.js
@@ -1,5 +1,5 @@
 // #region Imports
-import dbClient from "$lib/server/database/prisma/prisma.js"
+import dbClient from "$lib/server/database/prisma/dbClient.js"
 import logError from "$lib/server/utils/errorLogger.js"
 // #endregion
