+++ src/routes/(main)/(public)/reset/+page.server.js
@@ -1,5 +1,5 @@
 // #region Imports
-import dbClient from "$lib/server/database/prisma/prisma.js"
+import dbClient from "$lib/server/database/prisma/dbClient.js"
 import inputHandler from "$lib/server/utils/inputHandler.js"
 import { dateFromNow } from "$lib/client/utils/dateUtils.js"
 import { stringHasher } from "$lib/server/utils/hashUtils.js"
